Support max charge levels for the battery (do not show Charging when not)
Since recently there are settings to set "max charge level" for the
battery to stop charging before 100%, at least in the SteamDeck.

Due to vpower's logic, until now the interface was displaying info as
if it was charging undefinitely, but with percentages not moving and
time estimates being unknown/infinite, even if the hardware itself was
not charging the battery.

With this change, consider the battery to be "Full" when reaching this
maximum charge level, so the interface behaves accordingly.

Also, the time estimates for time until battery full are calculated
having this charge max level into account.