Have a default value when current/power files cannot be read
During tests, in rare occasions these messages were observed:

  [... several repeats of the vpower messages in short sequence ...]
  vpower[5099]: read /sys/class/power_supply/BAT1/current_now: No such device (os error 19)
  vpower[5099]: thread 'main' panicked at main.rs:294:67:
  vpower[5099]: Error: Missing necessary data: power_now_from_file
  vpower[5099]: note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
  systemd[1]: vpower.service: Main process exited, code=exited, status=101/n/a
  systemd[1]: vpower.service: Failed with result 'exit-code'.
  systemd[1]: vpower.service: Scheduled restart job, restart counter is at 5.
  steamdeck-oled systemd[1]: Started vpower daemon.
  vpower[5104]: Found AC power supply: '/sys/class/power_supply/ACAD'
  vpower[5104]: Found battery: /sys/class/power_supply/BAT1
  vpower[5104]: request_shutdown_battery_percent: 0.49999998
  vpower[5104]: force_shutdown_timeout_secs: 10
  vpower[5104]: Using sensor: steamdeck_hwmon
  vpower[5104]: Running.
  vpower[5104]: read /sys/class/power_supply/BAT1/current_now: No such device (os error 19)
  vpower[5104]: thread 'main' panicked at main.rs:294:67:
  vpower[5104]: Error: Missing necessary data: power_now_from_file
  vpower[5104]: note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
  systemd[1]: vpower.service: Main process exited, code=exited, status=101/n/a
  systemd[1]: vpower.service: Failed with result 'exit-code'.
  systemd[1]: vpower.service: Scheduled restart job, restart counter is at 6.
  systemd[1]: vpower.service: Start request repeated too quickly.
  systemd[1]: vpower.service: Failed with result 'exit-code'.
  systemd[1]: Failed to start vpower daemon.

Apparently, sometimes these battery/power devices "disappear" from the
OS point of view, seems somehow related with suspending, but the
actual problem has not been pinpointed yet.

With the previous code, vpower will abort since it expects either
'current_now' or 'power_now' to be available and be able to read data
from them.  And if aborting a few times in a row, systemd will stop
the service permanently, which is bad.

However, under these circumstances where data just is not available,
vpower is a bit powerless to do anything meaningful about it and
aborting doesn't really solve the problem, so just assume a value of
'0.0' and carry on the best that it can.