Read MaxChargeLevel from a file (written by helper) as D-Bus is not working
Currently steamos-manager has this property in the user-session, and
vpower running as root (which needs to perform one important function,
the poweroff) cannot communicate with it.

So instead, have a helper to read the value, and vpower just reads the
value as if from other files.

This value is needed mostly to calculate more accurately the time
until "full" (which, since recently, means "reached the desired charge
level"), as well as reaching Full status.

Also change the value when considered full, because 1% is too wide a
margin (and it keeps charging for half a minute or so, typically).