Read MaxChargeLevel from a /sys file if avail., as D-Bus is not working
Currently steamos-manager exposes this property in the user-session,
and vpower running as root (which needs to perform one important
function, the poweroff) cannot communicate with it easily and without
adding some complexity.

So instead, read a value directly from the file if available, as
already done from other files.  (Currently, steamos-manager does the
same, so if a particular system does not have those /sys files,
steamos-manager will not have more useful value than the default).

This value is needed mostly to calculate more accurately the time
until "full" (which, since recently, means "reached the desired charge
level"), as well as reaching Full status.

Also change the value when considered full, because 1% is too wide a
margin (and it keeps charging for half a minute or so, typically).