Add watchdog (binary and systemd service file)
This will help detect if vpower gets stuck, as there are reports in
the field about stuck battery percentage that suggest that it does.

If it does, one possible cause are problems reading files from /sys,
which seem to disappear in some very minimal fraction but still hit
now and then, and specially after suspend.

So this service does kill vpower in these situations, so its service
restarts and hopefully gets into a clean state.